requirements.txt
streamlit==1.35.0
supabase==2.0.3
python-dotenv==1.0.0
openai==1.12.0
numpy==1.26.2
pandas==2.1.4
plotly==5.19.0
groq==0.4.0 